totem/python/pyproject.toml
[tool.poetry]
name = "totem"
version = "0.1.0"
description = "Hardware control application for Raspberry Pi"
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "managers" }, 
    { include = "devices" },
    { include = "utils" },
    { include = "service" },
    { include = "tests" }
]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
pillow = ">=9.2.0,<12.0.0"
numpy = ">=1.23.0"
fastapi = "^0.110.0"
uvicorn = "^0.30.0"
websockets = "^12.0.0"
pydantic = "^2.6.0"
python-multipart = "^0.0.9"

# Optional dependencies for Raspberry Pi hardware
RPi-GPIO = { version = "^0.7.1", optional = true, markers = "sys_platform == 'linux'" }
spidev = { version = "^3.6", optional = true, markers = "sys_platform == 'linux'" }
gpiozero = { version = "^2.0", optional = true, markers = "sys_platform == 'linux'" }

eink-device = { path = "./devices/eink" }
nfc-device = { path = "./devices/nfc" }
wifi-device = { path = "./devices/wifi" }
nvme-device = { path = "./devices/nvme" }

[tool.poetry.extras]
raspberry-pi = ["RPi-GPIO", "spidev", "gpiozero"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# System tests
system-test = "tests.system_test:main"
serve = "service.main:start_server"

# E-ink service
eink-service = "scripts.start_eink_service:main"
eink-service-stop = "scripts.stop_eink_service:main"
eink-service-status = "scripts.eink_service_status:main"

# E-ink display tests
eink-debug = "tests.eink.test_eink_debug:main"
eink-pattern = "tests.eink.test_eink_pattern:main"
eink-quick-test = "tests.eink.eink_quick_test:main"
eink-diag = "tests.eink.eink_diag:main"
eink-gpio-debug = "tests.eink.eink_gpio_debug:main"
eink-pi5-test = "tests.eink.eink_pi5_test:main"
pi5-eink-test = "tests.eink.test_pi5_eink:main"
eink-sw-cs-test = "tests.eink.test_eink_sw_cs:main"
eink-nvme-conflict-test = "tests.eink.eink_nvme_conflict_test:main"
eink-2in13-test = "tests.eink.test_2in13_eink:main"
eink-spi-test = "tests.eink.eink_spi_test:main"
eink-spi-alt-test = "tests.eink.eink_spi_alt_test:main"
test-exact-driver = "tests.eink.test_exact_driver:main"
nfc-pn532-test = "tests.test_nfc_PNC532:main"

# Storage tests
nvme-test = "tests.test_nvme:main"
storage-test = "tests.test_storage_manager:main"
format-nvme = "scripts.format_nvme:main"

# GPIO tests
gpio-test = "tests.simple_gpio_test:main"
gpio-cleanup = "scripts.gpio_cleanup:main"

# Hardware diagnostics
hardware-check = "scripts.hardware_check:main"

# Scripts for setting up systems
setup-wifi-hotspot = "scripts.setup_wifi_hotspot:main"
setup-pi-system = "scripts.setup_pi_system:main"
setup-network-routing = "scripts.setup_network_routing:main"
setup-dependencies = "scripts.setup_dependencies:main"

# Deployment and installation scripts
install-totem = "scripts.install_totem:main"
deploy = "scripts.deploy:main"

# Diagnostic and fix scripts
fix-eink = "scripts.fix_eink:main"
eink-diagnostics = "scripts.eink_diagnostics:main"
fix-eink-dependencies = "scripts.fix_eink_dependencies:main"